Completed the Data Validation and replaced phone type with book type
SInce the supplier's phone number is nullable, a phone type will not always need to have a value if a book does not have a supplier info. However, a book in the store I am building can either be one of the following: hardcover, paperback or electronic. I refactored this logic in the code.